Need to exclude MyDerivedClassWithRowVersion from tests
