+++ src/NServiceBus.NHibernate.Tests/NServiceBus.NHibernate.Tests.csproj
@@ -73,6 +73,7 @@
     <Compile Include="SagaPersister\FakeConfigurationSource.cs" />
     <Compile Include="SagaPersister\InMemoryFixture.cs" />
     <Compile Include="SagaPersister\MySaga.cs" />
+    <Compile Include="SagaPersister\SessionFactoryHelper.cs" />
     <Compile Include="SagaPersister\SQLiteConfiguration.cs" />
     <Compile Include="SagaPersister\TestSaga.cs" />
     <Compile Include="SagaPersister\When_automapping_sagas.cs" />
