+++ src/NServiceBus.NHibernate.Tests/SagaPersister/When_building_the_sessionfactory.cs
@@ -1,12 +1,7 @@
 namespace NServiceBus.SagaPersisters.NHibernate.Tests
 {
     using System.Collections.Generic;
-    using System.Linq;
-    using Config.Internal;
-    using global::NHibernate.Cfg;
-    using global::NHibernate.Impl;
     using NUnit.Framework;
-    using Saga;
 
     [TestFixture]
     public class When_building_the_sessionFactory
@@ -16,13 +11,7 @@ namespace NServiceBus.SagaPersisters.NHibernate.Tests
         [Test]
         public void References_of_the_persistent_entity_should_also_be_mapped()
         {
-            var assemblyContainingSagas = typeof (TestSaga).Assembly;
-            var types = assemblyContainingSagas.GetTypes().ToList();
-            types.Add(typeof(ContainSagaData));
-
-            var builder = new SessionFactoryBuilder(types);
-
-            var sessionFactory = builder.Build(new Configuration().AddProperties(testProperties)) as SessionFactoryImpl;
+            var sessionFactory = SessionFactoryHelper.Build();
 
             Assert.NotNull(sessionFactory.GetEntityPersister(typeof (RelatedClass).FullName));
         }
